// src/platform/plugins/shared/visualizations/public/legacy/embeddable/i_embeddable.d.ts
export type EmbeddableError = ErrorLike;
export interface EmbeddableInput {
    version?: string;
    viewMode?: ViewMode;
    title?: string;
    description?: string;
    /**
     * Note this is not a saved object id. It is used to uniquely identify this
     * Embeddable instance from others (e.g. inside a container).  It's possible to
     * have two Embeddables where everything else is the same but the id.
     */
    id: string;
    lastReloadRequestTime?: number;
    hidePanelTitles?: boolean;
    /**
     * Reserved key for enhancements added by other plugins.
     */
    enhancements?: SerializableRecord;
    /**
     * List of action IDs that this embeddable should not render.
     */
    disabledActions?: string[];
    /**
     * Whether this embeddable should not execute triggers.
     */
    disableTriggers?: boolean;
    /**
     * Search session id to group searches
     */
    searchSessionId?: string;
    /**
     * Flag whether colors should be synced with other panels
     */
    syncColors?: boolean;
    /**
     * Flag whether cursor should be synced with other panels on hover
     */
    syncCursor?: boolean;
    /**
     * Flag whether tooltips should be synced with other panels on hover
     */
    syncTooltips?: boolean;
    executionContext?: KibanaExecutionContext;
}
export interface EmbeddableOutput {
    loading?: boolean;
    rendered?: boolean;
    error?: EmbeddableError;
    editUrl?: string;
    editApp?: string;
    editPath?: string;
    defaultTitle?: string;
    defaultDescription?: string;
    title?: string;
    description?: string;
    editable?: boolean;
    inlineEditable?: boolean;
    editableWithExplicitInput?: boolean;
    savedObjectId?: string;
}